feat: increase resilience by isolating brokers
before this change, when the process was starting up and a single broker could not be reached, the program paniced. This is less than ideal if a broker is down or has an expired tls certificate, for instance. Now, if a broker can not be connected to when the program starts up, we continually retry to connect to the broker over time.